+++ src/main/java/com/sofodev/armorplus/ArmorPlus.java
@@ -78,7 +78,7 @@ public class ArmorPlus {
 
     public static final String MODID = "armorplus";
     public static final String MODNAME = "ArmorPlus";
-    public static final String VERSION = "1.16.5-16.5.2";
+    public static final String VERSION = "1.16.5-16.7.0";
     public static final Logger LOGGER = LogManager.getLogger(MODID);
     public static final PacketHandler PACKET_HANDLER = new PacketHandler();
     /**