+++ src/main/java/com/sofodev/armorplus/registry/ModEnchantments.java
@@ -1,9 +1,6 @@
 package com.sofodev.armorplus.registry;
 
-import com.sofodev.armorplus.registry.enchantments.FuriousEnchantment;
-import com.sofodev.armorplus.registry.enchantments.LifeStealEnchantment;
-import com.sofodev.armorplus.registry.enchantments.SoulStealerEnchantment;
-import com.sofodev.armorplus.registry.enchantments.UnknownEnchantment;
+import com.sofodev.armorplus.registry.enchantments.*;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.fml.common.Mod;
@@ -23,6 +20,7 @@ public class ModEnchantments {
     public static final RegistryObject<Enchantment> LIFE_STEAL = register("life_steal", LifeStealEnchantment::new);
     public static final RegistryObject<Enchantment> SOUL_STEALER = register("soul_stealer", SoulStealerEnchantment::new);
     public static final RegistryObject<Enchantment> UNKNOWN = register("unknown", UnknownEnchantment::new);
+    public static final RegistryObject<Enchantment> SOUL_HARDEN = register("soul_harden", SoulHardenEnchantment::new);
 
     public static RegistryObject<Enchantment> register(String name, Supplier<? extends Enchantment> sup) {
         return ENCHANTMENTS.register(name, sup);